+++ packages/web/components/index.ts
@@ -8,3 +8,6 @@ export { default as TaskMonitor } from './TaskMonitor';
 export { default as ControlPanel } from './ControlPanel';
 export { default as TerminalLivestream } from './TerminalLivestream';
 export { default as ToolCard } from './ToolCard';
+
+// Version badge
+export { default as VersionBadge } from './VersionBadge';